+++ src/main/java/fr/mrcraftcod/gunterdiscord/settings/configs/WarnTimeConfig.java
@@ -17,7 +17,7 @@ public class WarnTimeConfig extends ValueConfiguration
 	{
 		try
 		{
-			super.setValue(guild, Integer.parseInt(value.toString()));
+			super.setValue(guild, Double.parseDouble(value.toString()));
 		}
 		catch(Exception e)
 		{
